% Auto-generated: do not edit by hand
\name{bubble}

\alias{bubble}

\title{Bubble component}

\description{
Bubble component using d3 and hooks
}

\usage{
bubble(id=NULL, width=NULL, height=NULL, data=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{width}{Numeric. Chart width}

\item{height}{Numeric. Chart height}

\item{data}{Unnamed list. Data}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

